+++ core/src/main/java/org/apache/shiro/mgt/AbstractRememberMeManager.java
@@ -28,10 +28,10 @@ import org.apache.shiro.crypto.cipher.CipherService;
 import org.apache.shiro.lang.io.DefaultSerializer;
 import org.apache.shiro.lang.io.Serializer;
 import org.apache.shiro.lang.util.ByteSource;
+import org.apache.shiro.lang.util.ByteUtils;
 import org.apache.shiro.subject.PrincipalCollection;
 import org.apache.shiro.subject.Subject;
 import org.apache.shiro.subject.SubjectContext;
-import org.apache.shiro.util.ByteUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
