+++ crypto/cipher/src/test/groovy/org/apache/shiro/crypto/cipher/AesCipherServiceTest.groovy
@@ -22,7 +22,7 @@ package org.apache.shiro.crypto.cipher
 import org.apache.shiro.lang.codec.CodecSupport
 import org.apache.shiro.lang.util.ByteSource
 import org.apache.shiro.lang.util.Destroyable
-import org.apache.shiro.util.ByteUtils
+import org.apache.shiro.lang.util.ByteUtils
 import org.bouncycastle.jce.provider.BouncyCastleProvider
 import org.junit.Test
