+++ crypto/cipher/src/test/groovy/org/apache/shiro/crypto/cipher/BlowfishCipherServiceTest.groovy
@@ -21,7 +21,7 @@ package org.apache.shiro.crypto.cipher
 
 import org.apache.shiro.lang.codec.CodecSupport
 import org.apache.shiro.lang.util.ByteSource
-import org.apache.shiro.util.ByteUtils
+import org.apache.shiro.lang.util.ByteUtils
 import org.junit.Test
 
 import static org.junit.Assert.assertTrue
